Refactoring CEP application packages
